+++ dali/tests/test_command.py
@@ -1,41 +1,40 @@
+from __future__ import unicode_literals
+import os
+import sys
+import unittest
+
+
 try:
-    from dali import address
-    from dali import command
-    from dali import frame
-    from dali.gear import general as generalgear
-    from dali.gear import emergency
-    from dali.gear import incandescent
-    from dali.gear import led
-    from dali.device import general as generaldevice
+    import dali
 except ImportError:
     # Realign paths, and try import again
     # Since pyCharm's unittest runner fails on relative imports
-    import os
-    import sys
+    path = os.path
+    PACKAGE_PARENT = '../..'
+    SCRIPT_DIR = path.dirname(
+        path.realpath(path.join(os.getcwd(), path.expanduser(__file__)))
+    )
+    sys.path.append(path.normpath(path.join(SCRIPT_DIR, PACKAGE_PARENT)))
 
-    PACKAGE_PARENT = '..'
-    SCRIPT_DIR = os.path.dirname(
-        os.path.realpath(os.path.join(os.getcwd(),
-                         os.path.expanduser(__file__))))
-    sys.path.append(os.path.normpath(os.path.join(SCRIPT_DIR, PACKAGE_PARENT)))
 
-    from dali import address
-    from dali import command
-    from dali import frame
-    from dali.gear import general as generalgear
-    from dali.gear import emergency
-    from dali.gear import incandescent
-    from dali.gear import led
-    from dali.device import general as generaldevice
-import unittest
-import sys
+from dali import address
+from dali import command
+from dali import frame
+from dali.device import general as generaldevice
+from dali.gear import emergency
+from dali.gear import general as generalgear
+from dali.gear import incandescent
+from dali.gear import led
+
 
 if sys.version_info[0] == 3:
     xrange = range
 
+
 # Only test device types up to this number - if we went all the way to
 # 255 the tests would take unnecessarily long!
 max_devicetype = 10
+
 
 def _test_pattern():
     # Control gear, device type zero
@@ -59,30 +58,33 @@ def _test_pattern():
     for c in (0xc5, 0xc7, 0xc9):
         yield 24, (c, 0x00, 0x00), 0
 
+
 class TestCommands(unittest.TestCase):
 
     def test_test_coverage(self):
-        "all command classes are covered by test pattern"
+        """all command classes are covered by test pattern"""
         seen = {}
         for fs, d, dt in _test_pattern():
             c = command.from_frame(frame.ForwardFrame(fs, d), devicetype=dt)
             seen[c.__class__] = True
         for cls in command.Command._commands:
-            self.assertTrue(cls in seen,
-                            "class {} not covered by tests".format(cls.__name__))
+            self.assertTrue(
+                cls in seen,
+                'class {} not covered by tests'.format(cls.__name__)
+            )
 
     def test_roundtrip(self):
-        "all frames survive command.from_frame()"
+        """all frames survive command.from_frame()"""
         for fs, d, dt in _test_pattern():
             f = frame.ForwardFrame(fs, d)
             c = command.from_frame(f, dt)
             nf = c.frame
             self.assertEqual(
-                f, nf, "frame {} failed command round-trip; command {} "
-                "became {}".format(str(f), str(c), str(nf)))
+                f, nf, 'frame {} failed command round-trip; command {} '
+                'became {}'.format(str(f), str(c), str(nf)))
 
     def test_unicode(self):
-        "command objects return unicode from their __unicode__ method"
+        """command objects return unicode from their __unicode__ method"""
         # Python 2 only
         if sys.version_info[0] == 2:
             for fs, d, dt in _test_pattern():
@@ -91,32 +93,41 @@ class TestCommands(unittest.TestCase):
                                       unicode)
 
     def test_str(self):
-        "command objects can be converted to strings"
+        """command objects can be converted to strings"""
         for fs, d, dt in _test_pattern():
             f = frame.ForwardFrame(fs, d)
             self.assertIsInstance(str(command.from_frame(f, dt)),
                                   str)
 
     def test_with_integer_destination(self):
-        "commands accept integer destination"
-        self.assertEqual(generalgear.DAPC(5, 100).destination, address.Short(5))
+        """commands accept integer destination"""
+        self.assertEqual(
+            generalgear.DAPC(5, 100).destination,
+            address.Short(5)
+        )
         self.assertEqual(generalgear.Off(5).destination, address.Short(5))
         self.assertRaises(ValueError, generalgear.Off, -1)
         self.assertRaises(ValueError, generalgear.Off, 64)
         self.assertRaises(ValueError, generalgear.Off, None)
 
     def test_response(self):
-        "responses act sensibly"
+        """responses act sensibly"""
         for fs, d, dt in _test_pattern():
             f = frame.ForwardFrame(fs, d)
             c = command.from_frame(f, dt)
             if c._response:
                 if sys.version_info[0] == 2:
-                    self.assertIsInstance(c._response(None).__unicode__(), unicode)
-                self.assertRaises(TypeError, lambda: c._response("wibble"))
+                    self.assertIsInstance(
+                        c._response(None).__unicode__(),
+                        unicode
+                    )
+                self.assertRaises(TypeError, lambda: c._response('wibble'))
                 if sys.version_info[0] == 2:
-                    self.assertIsInstance(c._response(frame.BackwardFrame(0xff))\
-                                          .__unicode__(), unicode)
+                    self.assertIsInstance(
+                        c._response(frame.BackwardFrame(0xff)).__unicode__(),
+                        unicode
+                    )
 
-if __name__ == "__main__":
+
+if __name__ == '__main__':
     unittest.main()